@@ -4,7 +4,7 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "8.0.* || 8.1.*",
+        "php": "8.1.* || 8.1.*",
         "ext-json": "*",
         "phpmailer/phpmailer": "^6.6",
         "predis/predis": "^2.2",
